Move Floats.java and Doubles.java to javasupport/expressions
